identity contracts split into parts: contact itself, read-only interface for other contracts, common lib #2562
