+++ contracts/identity/identity.cpp
@@ -1,8 +1,7 @@
-#include <eosiolib/chain.h>
+#include "common.hpp"
+
 #include <eosiolib/contract.hpp>
 #include <eosiolib/dispatcher.hpp>
-#include <eosiolib/singleton.hpp>
-#include <eosiolib/multi_index.hpp>
 #include <eosiolib/vector.hpp>
 
 namespace identity {
@@ -49,169 +48,10 @@ namespace identity {
     *
     *  When an account certs a property we check to see if that
     */
-   class contract : public eosio::contract {
+   class contract : public identity_base {
       public:
 
-         using eosio::contract::contract;
-
-         typedef uint64_t identity_name;
-         typedef uint64_t property_name;
-         typedef uint64_t property_type_name;
-
-         struct certvalue {
-            property_name property; ///< name of property, base32 encoded i64
-            string     type; ///< defines type serialized in data
-            vector<char>  data; ///<
-            string        memo; ///< meta data documenting basis of certification
-            uint8_t       confidence = 1; ///< used to define liability for lies,
-                                          /// 0 to delete
-
-            EOSLIB_SERIALIZE( certvalue, (property)(type)(data)(memo)(confidence) )
-         };
-
-         struct certrow {
-            uint64_t            id;
-            property_name       property;
-            uint64_t            trusted;
-            account_name        certifier;
-            uint8_t             confidence = 0;
-            string              type;
-            vector<char>        data;
-            uint64_t primary_key() const { return id; }
-            /* constexpr */ static key256 key(uint64_t property, uint64_t trusted, uint64_t certifier) {
-               /*
-               key256 key;
-               key.uint64s[0] = property;
-               key.uint64s[1] = trusted;
-               key.uint64s[2] = certifier;
-               key.uint64s[3] = 0;
-               */
-               return key256::make_from_word_sequence<uint64_t>(property, trusted, certifier);
-            }
-            key256 get_key() const { return key(property, trusted, certifier); }
-
-            EOSLIB_SERIALIZE( certrow , (property)(trusted)(certifier)(confidence)(type)(data)(id) )
-         };
-
-         struct identrow {
-            uint64_t     identity;
-            account_name creator;
-
-            uint64_t primary_key() const { return identity; }
-
-            EOSLIB_SERIALIZE( identrow , (identity)(creator) )
-         };
-
-         struct trustrow {
-            account_name account;
-
-            uint64_t primary_key() const { return account; }
-
-            EOSLIB_SERIALIZE( trustrow, (account) )
-         };
-
-         typedef eosio::multi_index<N(certs), certrow,
-                                    eosio::indexed_by< N(bytuple), eosio::const_mem_fun<certrow, key256, &certrow::get_key> >
-                                    > certs_table;
-         typedef eosio::multi_index<N(ident), identrow> idents_table;
-         typedef singleton<N(account), identity_name>  accounts_table;
-         typedef eosio::multi_index<N(trust), trustrow> trust_table;
-
-         identity_name get_claimed_identity( account_name acnt ) {
-            return accounts_table( _self, acnt ).get_or_default(0);
-         }
-
-         account_name get_owner_for_identity( uint64_t receiver, identity_name ident ) {
-            // for each trusted owner certification
-            //   check to see if the certification is still trusted
-            //   check to see if the account has claimed it
-            certs_table certs( _self, ident );
-            auto idx = certs.template get_index<N(bytuple)>();
-            auto itr = idx.lower_bound(certrow::key(N(owner), 1, 0));
-            account_name owner = 0;
-            while (itr != idx.end() && itr->property == N(owner) && itr->trusted) {
-               if (sizeof(account_name) == itr->data.size()) {
-                  account_name account = *reinterpret_cast<const account_name*>(itr->data.data());
-                  if (ident == get_claimed_identity(account)) {
-                     if (is_trusted(itr->certifier) ) {
-                        // the certifier is still trusted
-                        if (!owner || owner == account) {
-                           owner = account;
-                        } else {
-                           //contradiction found: different owners certified for the same identity
-                           return 0;
-                        }
-                     } else if (_self == receiver){
-                        //the certifier is no longer trusted, need to unset the flag
-                        idx.modify(itr, 0, [&](certrow& r) {
-                              r.trusted = 0;
-                           });
-                     } else {
-                        // the certifier is no longer trusted, but the code runs in read-only mode
-                     }
-                  }
-               } else {
-                  // bad row - skip it
-               }
-               ++itr;
-            }
-            if (owner) {
-               //owner found, no contradictions among certifications flaged as trusted
-               return owner;
-            }
-            // trusted certification not found
-            // let's see if any untrusted certifications became trusted
-            itr = idx.lower_bound(certrow::key(N(owner), 0, 0));
-            while (itr != idx.end() && itr->property == N(owner) && !itr->trusted) {
-               if (sizeof(account_name) == itr->data.size()) {
-                  account_name account = *reinterpret_cast<const account_name*>(itr->data.data());
-                  if (ident == get_claimed_identity(account) && is_trusted(itr->certifier)) {
-                     if (_self == receiver) {
-                        // the certifier became trusted and we have permissions to update the flag
-                        idx.modify(itr, 0, [&](certrow& r) {
-                              r.trusted = 1;
-                           });
-                     }
-                     if (!owner || owner == account) {
-                        owner = account;
-                     } else {
-                        //contradiction found: different owners certified for the same identity
-                        return 0;
-                     }
-                  }
-               } else {
-                  // bad row - skip it
-               }
-               ++itr;
-            }
-            return owner;
-         }
-
-         identity_name get_identity_for_account( uint64_t receiver, account_name acnt ) {
-            //  check what identity the account has self certified owner
-            //  verify that a trusted certifier has confirmed owner
-            auto identity = get_claimed_identity(acnt);
-            return (identity != 0 && acnt == get_owner_for_identity(receiver, identity)) ? identity : 0;
-         }
-
-         bool is_trusted_by( account_name trusted, account_name by ) {
-            trust_table t( _self, by );
-            return t.find( trusted ) != t.end();
-         }
-
-         bool is_trusted( account_name acnt ) {
-            account_name active_producers[21];
-            auto count = get_active_producers( active_producers, sizeof(active_producers) );
-            for( size_t i = 0; i < count; ++i ) {
-               if( active_producers[i] == acnt )
-                  return true;
-            }
-            for( size_t i = 0; i < count; ++i ) {
-               if( is_trusted_by( acnt, active_producers[i] ) )
-                  return true;
-            }
-            return false;
-         }
+        using identity_base::identity_base;
 
          void settrust( const account_name trustor, ///< the account authorizing the trust
                         const account_name trusting, ///< the account receiving the trust
