+++ contracts/identity/test/CMakeLists.txt
@@ -2,6 +2,6 @@ set(ABI_FILES "identity_test.abi")
 configure_file("${ABI_FILES}" "${CMAKE_CURRENT_BINARY_DIR}" COPYONLY)
 add_wast_executable(TARGET identity_test
   INCLUDE_FOLDERS "${STANDARD_INCLUDE_FOLDERS}"
-  LIBRARIES libc++ libc eosiolib
+  LIBRARIES identity_interface identity_common libc++ libc eosiolib
   DESTINATION_FOLDER ${CMAKE_CURRENT_BINARY_DIR}
 )
